single python code through minimal set of macros
